# Faster editor line drawing - Path2D and draw_line