@lawnjelly 2021-11-10
Just to note, got some further testing of the line drawing side of this in #54826. When trying that test project:

* Drawing lots of thick lines is much faster, as expected due to batching (from 145fps to 870fps)
* Drawing without batching is actually a bit slower (255fps to 145fps), and this represents the difference between the line drawing routine and the poly drawing routine. This will also apply to single thick lines.

If anyone reports the latter as a problem, we can put in a check in the `canvas_item_add_line` routine to check whether batching is off, and revert to the old routine. It's comparable to the performance in 3.2.3 anyway.