@lawnjelly 2021-12-01
> This is not editor only, right? Eager to test this as we use custom path2Ds extensively.

This works outside the editor too.